requirements.txt
streamlit
python-dotenv
google-generativeai
sentence-transformers
numpy
pandas
scikit-learn
Pillow
